Invulnerability and stamina regeneration.

+++ Assets/Scripts/Data/Events/GameEvents/StaminaChanged.cs
@@ -2,8 +2,8 @@ namespace SoulsLike
 {
 	public readonly struct StaminaChanged : IGameEvent
 	{
-		public readonly int currentStamina;
+		public readonly float currentStamina;
 
-		public StaminaChanged(int stamina) => currentStamina = stamina;
+		public StaminaChanged(float stamina) => currentStamina = stamina;
 	}
 }
